refactor(eve): rename codex model directory to chatgpt
