+++ packages/eve/src/public/models/openai/chatgpt/auth.ts
