+++ packages/eve/src/public/models/openai/chatgpt/model.ts
